% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Create a linreg object}
\arguments{
\item{formula}{formula for linear model}

\item{data}{the dataset provided}
}
\value{
the output from the \code{\link{linreg}}
\dontrun{
linreg("hello")
}
}
\description{
Create a linreg object
}
\section{Fields}{

\describe{
\item{\code{formula}}{.}

\item{\code{data}}{.}

\item{\code{X}}{.}

\item{\code{Y}}{.}

\item{\code{beta_estimates}}{.}

\item{\code{esti_y}}{.}

\item{\code{residual}}{.}

\item{\code{t_params}}{.}

\item{\code{deg_freedom}}{.}

\item{\code{residualt}}{.}

\item{\code{resid_var}}{.}

\item{\code{residualstd}}{.}

\item{\code{betavariance}}{.}

\item{\code{bb}}{.}

\item{\code{tvalues}}{.}

\item{\code{pvalues}}{.}

\item{\code{standardizedresiduals}}{.}

\item{\code{sqrtstresiduals}}{.}

\item{\code{export_formula}}{.}

\item{\code{export_data}}{.}
}}

\section{Methods}{

\describe{
\item{\code{print()}}{Prints Model information}
}}

